data_insights_team/requirements.txt
agno
pandas
matplotlib
openai
python-dotenv
